service: hid: Split hid.cpp into individual interfaces
