+++ src/core/hid/input_interpreter.cpp
@@ -5,13 +5,14 @@
 #include "core/hid/hid_types.h"
 #include "core/hid/input_interpreter.h"
 #include "core/hle/service/hid/controllers/npad.h"
-#include "core/hle/service/hid/hid.h"
+#include "core/hle/service/hid/hid_server.h"
+#include "core/hle/service/hid/resource_manager.h"
 #include "core/hle/service/sm/sm.h"
 
 InputInterpreter::InputInterpreter(Core::System& system)
     : npad{system.ServiceManager()
-               .GetService<Service::HID::Hid>("hid")
-               ->GetAppletResource()
+               .GetService<Service::HID::IHidServer>("hid")
+               ->GetResourceManager()
                ->GetController<Service::HID::Controller_NPad>(Service::HID::HidController::NPad)} {
     ResetButtonStates();
 }
