+++ src/core/hle/service/hid/irs.h
@@ -3,14 +3,11 @@
 
 #pragma once
 
+#include "core/core.h"
 #include "core/hid/hid_types.h"
 #include "core/hid/irs_types.h"
 #include "core/hle/service/hid/irsensor/processor_base.h"
 #include "core/hle/service/service.h"
-
-namespace Core {
-class System;
-}
 
 namespace Core::HID {
 class EmulatedController;
